Fix API handlers for film details and weekly schedule
